refactor: address PR 196 review feedback on position recipients
- BasePositionRecipient no longer inherits TimelockedPositionRecipient;
  it owns positionManager and inherits ReentrancyGuardTransient directly.
  Recipient constructors drop the unused operator/timelock params.
- Amounts packed into a single slot (uint128 pair); onAmountsReceived and
  claim read the slot once.
- BeneficiaryVault.registerBeneficiary explicitly rejects the zero address.
- BuybackAndBurnPositionRecipient docs: caller approval requirement and
  why the native-currency0 check also fixes the burn target.

+++ src/interfaces/IClaimablePositionRecipient.sol
@@ -30,7 +30,7 @@ interface IClaimablePositionRecipient {
     function claim(uint256 tokenId, uint256 minCurrency0Amount, uint256 minCurrency1Amount) external;
 
     /// @notice Returns the amounts attributed to a position and available to claim
-    function amounts(uint256 tokenId) external view returns (uint256 currency0Amount, uint256 currency1Amount);
+    function amounts(uint256 tokenId) external view returns (uint128 currency0Amount, uint128 currency1Amount);
 
     /// @notice MUST be called after transferring amounts in to register the new balance. Failure to
     ///         do so will result in a loss of funds attribution.
